# Copy the baseVM in a goroutine and obtain via a channel