--- a/server/main.go
+++ b/server/main.go
@@ -27,6 +27,12 @@ func logDuration(s string) func() {
 	}
 }
 
+func vmCopier(baseVM *otto.Otto, c chan *otto.Otto) {
+	for {
+		c <- baseVM.Copy()
+	}
+}
+
 func main() {
 	logrus.StandardLogger().Level = logrus.DebugLevel
 
@@ -68,6 +74,9 @@ func main() {
 	}
 	jsInitialised()
 
+	vmChan := make(chan *otto.Otto)
+	go vmCopier(baseVM, vmChan)
+
 	tpl, err := template.ParseFiles("./index.template")
 	if err != nil {
 		panic(err)
@@ -93,7 +102,7 @@ func main() {
 
 	m.PathPrefix("/").HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		vmCopied := logDuration("copy vm")
-		vm := baseVM.Copy()
+		vm := <-vmChan
 		vmCopied()
 
 		l := loop.New(vm)