From your talk you were weren't super happy about the precious time wasted to copy the `baseVM` during a user's request.

This copies the `baseVM` in another goroutine so that when a request comes in the request's VM is already warm and waiting to be used. I haven't actually made the copying of the `baseVM` faster, just faster for obtaining one in a request.

# Results

Running on my laptop

Before:

```
DEBU[0027] copy vm [before]
DEBU[0027] copy vm [after]                               duration=12.043119ms
...
DEBU[0057] copy vm [before]
DEBU[0057] copy vm [after]                               duration=16.164511ms
...
DEBU[0057] copy vm [before]
DEBU[0057] copy vm [after]                               duration=12.105417ms
```

After:

```
DEBU[0036] copy vm [before]
DEBU[0036] copy vm [after]                               duration=15.606µs
...
DEBU[0038] copy vm [before]
DEBU[0038] copy vm [after]                               duration=14.46µs
...
DEBU[0039] copy vm [before]
DEBU[0039] copy vm [after]                               duration=14.94µs
```

So from ~15ms to ~15µs. 

To better handle bursts of user activity it might be a good idea to make it a buffered channel, so you can have a few otto VMs on standby instead of just one.