@deoxxa 2015-11-11
This is a great idea! I rebased this and merged it manually - thanks :)